requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.2
exceptiongroup==1.3.1
faiss-cpu==1.13.2
fastapi==0.136.0
filelock==3.28.0
fsspec==2026.3.0
ftfy==6.3.1
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==26.1
pika==1.3.2
pillow==12.2.0
psycopg2-binary==2.9.9
pydantic==2.13.1
pydantic_core==2.46.1
python-multipart==0.0.26
regex==2026.4.4
starlette==1.0.0
sympy==1.14.0
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
wcwidth==0.6.0
git+https://github.com/openai/CLIP.git
